# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Matcherino Draft Viewer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tournaments:
    get:
      operationId: listTournaments
      tags: [tournaments]
      summary: List all tracked tournaments
      responses:
        "200":
          description: List of tournaments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tournament"
    post:
      operationId: addTournament
      tags: [tournaments]
      summary: Add a Matcherino tournament by URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TournamentInput"
      responses:
        "201":
          description: Tournament added and data fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
        "400":
          description: Invalid URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Tournament already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
  /tournaments/{id}:
    get:
      operationId: getTournament
      tags: [tournaments]
      summary: Get a tournament by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tournament data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTournament
      tags: [tournaments]
      summary: Delete a tournament and its matches
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /tournaments/{id}/refresh:
    post:
      operationId: refreshTournament
      tags: [tournaments]
      summary: Re-fetch tournament data from Matcherino
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Refreshed tournament
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    TournamentInput:
      type: object
      properties:
        url:
          type: string
          description: Full Matcherino tournament URL (e.g. https://matcherino.com/t/bbl_s9_q3/overview)
      required:
        - url
    Tournament:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        url:
          type: string
        gameName:
          type: ["string", "null"]
        gameMode:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        matchCount:
          type: integer
        status:
          type: string
          description: fetching | done | error
        errorMessage:
          type: ["string", "null"]
          description: Set when status is error
        eventDate:
          type: ["string", "null"]
          format: date-time
          description: When the event took place (from Matcherino startAt)
        source:
          type: string
          description: manual | emea_auto
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - name
        - url
        - matchCount
        - status
        - source
        - createdAt
        - updatedAt
...